Moved a lock to a point after a trivial test
The lock in question doesn't need to synchronize access to CanAccept, which is already an atomic boolean.  There's no state here to become inconsistent, and most clients do not attempt to lock the dispatch lock before accepting dispatch delivery.  If we wind up racing, it will be to accept dispatch delivery, and at that point we either get it or we don't.