Altering tests to make use of a stronger ownership constraint and a weaker termination criteria
There are several cases that are implied by the previous versions of these tests which are incompatible with each other:

1)  CoreThread must function properly even if it has the final unique shared_ptr to its own enclosing context
2)  Client expects CoreThread Wait to return only after releasing its enclosing context
3)  SimpleOwnershipValidator expects CoreContext members to be destroyed before CoreContext

Constraint 1 can only be satisfied if CoreThread is not destroyed when it's time to signal waiters.  Constraint 2 requires that the CoreThread release all references to its enclosing context, which means that it must hold a self reference in order to prevent premature destruction.  Constraint 3, however, prevents this exact strategy by requiring that all references to context members be released prior to context destruction.

Arguably, a way out is based on the observation that constraint 1 becomes irrelevant if if there are any waiters, because such waiters posses a shared pointer to the CoreThread and therefore must also posses a shared pointer to the enclosing context in order to avoid running afoul of constraint 3.  Unfortunately this requires that the signaling of the termination condition be conditionally dependent on the uniqueness of the context reference, and because one can only guarantee correctness for one side of this branch (IE, if a shared_ptr's use count is one, it can never unexpectedly change, but if it is anything else, it could unexpectedly become one) such a behavior cannot be implemented in the general case.  The alternative solution, that signaling the termination condition be dependent on the existence of waiters, would either require the use of an additional synchronization primitive and merely defer the problem, or would require the violation of Wait's idempotency properties.