Formalizing running context behavior to now be single-voter shutdown rather than consensus shutdown, unit tests updated
The original conception of a subcontext was based on the idea that a subcontext is a client to its parent context, which is the service provider.  This concept has been abandoned, and instead, child contexts are now subunits of work executed based on entities in the parent context.  Thus, votes to terminate a parent context must percolate unconditionally downwards to the child contexts.