Adding stronger guarantees of object cleanup during rundown, moving dispatch queue exhaustion behavior to CoreThread::DoRun
During rundown, exceptions are thrown which unwind and cause threads to enter a stopped state.  In this state, formerly, the thread's dispatch queue length may potentially be nonzero, but this is fine because the queue is eventually cleaned up with the thread's termination.  If, however, one of the pended events contains a reference back to its owning thread, then the existence of this entry in the dispatch queue creates a cyclic reference that will not be cleaned up automatically.

A new guarantee is therefore required that all threads in a stopped state have a dispatch queue length of precisely zero.