WinUSB preallocates URB spaces before performing UVC commit #9239
Formerly, all allocations were taking place in the QueueAllOverlappedReads method.  This was causing a huge delay between when the commit takes place and when the first URB is received by the device, which causes the device to believe that it is streaming too quickly for the host and to initiate backoff.  Preallocating the necessary buffers shortens this time, and allows for a much shorter interval between commit and initial read.

Conflicts:
	source/Autowiring/ObjectPool.h