Making the ObjectPool a movable type, adding corresponding test
Adding some overrides that let an object pool be moved.  Keep in mind that, while this construction does allow you to move a pool even while objects are returning to it asynchronously, it's still generally unsafe to directly make use of a type which might be asynchronously moved like this.  You can move a pool if there are objects coming back, but not if there are threads trying to take objects out.