Descendants are now obtained using CoreContext proper, rather than trying to keep duplicates
This also corrects a race condition that arises whenever a context is being terminated at about the same time as a descendant context is being terminated.  In this situation, the contents of an AutoPacketFactory have already been unlinked and invalidated at the same time as that packet factory is being sent back up the stack to have its contents unlinked.

By centralizing an AutoPacketFactory so that its sole concern are the AutoFilters in the current context, and then using a context enumerator to find AutoFilters in descendant contexts, AutoPacketFactory now has no race condition over its own contents.  Descendant context destruction only results in the invalidation of a packet factory's object pool, which will trigger regeneration at a later time, rather than unregistration unwind as is currently required.