Adding pre-bolt injection support
The creator of a context should be able to obtain total prior control over what goes into a context in advance of any bolts that are fired.  Context creators should use this power very sparingly, as it does imply that there are some entities added to a context which cannot be overridden by bolts, but in the rare cases where this power is desired, it has been made available here.