Adding reset(target) and transfer(target) methods to atomic_object, and by extension to shared_object.
NOTE: It is not clear that these conditional transfer methods are fundamental to atomic_object. Instead, write a class that provides conditional assignment and transfer methods... which seems like a separate project.

NOTE: It will then be necessary to extend the unlock_object functionality to an atomic array of unlocks.