Adding a dacl to the server-side named pipe
A DACL, or Discretionary Access Control List, is required to allow a server-side named pipe to be read by a client.  We cannot simply permit "all access" in this case because this would grant untrusted users the ability to modify the DACL itself or to change its ownership, and this could cause bad behavior, so instead we allow all access only to members of the administrators group and permit read/write access to all other local users.

The IPC client was also made more resilient by adding the capability to detect when a named pipe read attempt failed due to permissions problems