De-templating DispatchQueue, adding static assertions to check time limits
DispatchQueue can only operate on resolutions of about a microsecond anyway due to limits on the scheduler, and it also really needs the clock type to be steady.  Because changes to the system time could seriously mess up the dispatch queue, we can explicitly use this type in the DispatchThunkDelayedExpression and eliminate the template.  We can also make a static assertion to ensure that passed times satisfy the microsecond resolution requirement.