Recommend do not override CMAKE_INSTALL_PREFIX on MSVC
I'm a Windows user, I don't actually want this--I can set CMAKE_INSTALL_PREFIX myself if necessary.  When I install websocketpp, I intend for the installation to be to a place that's similar in sense to what it is on Unix systems, which on Windows is generally Program Files.  By allowing the CMAKE_INSTALL_PREFIX path to simply default, not only does this default behavior automatically appear, but it also corrects a problem currently existing with the INSTALL project which causes the websocketpp-config.cmake installed by the INSTALL project to contain invalid entries.