AutowiringEnclosure MUST be header-only
This is needed in order to prevent near-incompatibility issues from causing programs to crash when trying to link against the AutowiringEnclosure.  The root problem, here, is Google's recommendation that users distribute googletest in their own applications.  This means that there is no public-facing API that Google is compelled to maintain (in order to evade linker errors) and therefore no ABI to standardize behavior.

So, in order to ensure that our class receives the correct layout, we have to make it the consumer's responsibility to ensure correct include ordering.