Cleaning up boost find_package and linkage spec
We shouldn't be attempting to link with pthread directly; Boost provides a find_package for pthread that we should use to assess whether we should link to it.

If we're going to reference Boost with find_package, we should probably reference all of the components we're using in one go.