Adding unit test for desired last-call behavior.
An argument of type AutoPacket& is satisfied immediately and can be used to add data to a packet immediately. Likewise, const AutoPacket& cannot add data, and is therefore understood to be a last-call. If there are additional arguments then the call will be made only if those arguments can also be satisfied.