Pipe & Broadcast declarations Ignore AutoPacket type, since it is always broadcast.
When resolving all consistent types to pipe for A -> B, an AutoPacket argument in B indicates that all types sent by A might be referenced. Likewise, and AutoPacket argument in A means that all types requested by B might be supplied.

This enables the deduction of pipes from an initial call A::AutoFilter(AutoFilter&) or to a final call B::AutoFilter(const AutoFilter&).