Implemented type independent extraction by AutoStile.
If no output arguments are provided it is assumed that all data in the slave context should be output.

This relies on collision avoidance by AutoPacket::ForwardAll. If the data in the slave context is already present in the master context (because it is the data that was previously injected) then ForwardAll will discard the data.