Status of LibCXX is now validated in version checker
Because autowiring_USE_LIBCXX affects whether or not the requested version is compilable by a client, we now allow a client to specify this flag before invoking find_package in order to assess compatibility of the client's request with the version of autowiring identified.  This is a necessary first step to getting side-by-side installations of Autowiring.