Deprecation of Stile and Pipe in preparation for a concept redux
Will be rolling Stile and Pipe into a new concept of generalized AutoPacket fan in/fan out.  This new concept will be optimized for speed and ease of use, but as long as Stile and Pipe exist, there is a significant amount of overlap and occlusion.  Thus, these features are being deprecated ahead of the new functionality.