Eliminating boost::coroutine
There's nothing strictly wrong with coroutines, but we currently have no use for them in Autowiring.  The current preference is to make use of OS-provided thread pooling features rather than making use of coroutines where they are used for parallelization, and in other cases where they are used to implement yield, we use explicit state.