Packet factory is no longer recursive
This is a breaking change, and comes with a bump to the version number.  The bottom line is that AutoFilters are no longer recursive with respect to contexts.  This change was necessary because, unlike events, AutoFilters operating under recursion can violate the confinement principle by attaching decorations.  This could cause a situation where two subcontexts interfere indirectly with each other due to attached decorations.

Rather than trying to provide a backwards-compatible path for this feature, the feature has instead been completely deprecated.  This feature's theory of operation will require closer examination and justification at a later time.