Deprecation of the Event Stream system
This system is being deprecated as it is ill-maintained, disused, and very out of date.  The concepts represented here must be reimplemented to support AutoPacket, and will almost certainly require the support of a serialization strategy such as Protobuf or CaptProto.

Until the concept can be revisited, this code should be eliminated in order to facilitate header load reduction.