Eliminating all external references to boost
This allows us to be a completely self-contained library build, and also corrects some build issues on Android.  Additionally, boost sources that must be built are now built as a part of the Autowiring project, rather than the AutoNet project, in order to make boost std C++11 compatibility types available to the autowiring project.