Adding a packet `Wait` routine
This routine allows consumers to block for a packet to contain a specific set of desired decorations.  This method will work even if processing on this packet crosses thread boundaries.  A `CallFast` variant is provided which provides the same services, but requires users to allocate storage for a `condition_variable` type, and also requires that all passed decorations specified as a `shared_ptr` type.