Fixing an invalid default construction bug
This bug attempted to default construct output arguments of the form `std::shared_ptr<T>&`.  The reason this output signature is supported is precisely to avoid requiring that Autowiring take responsibility for all construction events on `AutoPacket` decorations.  This bug has been fixed, and a test added to prevent future regressions.