Eliminate return type in cases where the address is implied
The only cases where we need to return a reference to the decorated type are cases where this information is not already available to the user.  Decorating with `std::shared_ptr` implies that the user already has access to the decorated type's memory (as they must have constructed it) making a return from this function redundant.  Furthermore, because we intend to support decoration with a null shared pointer, it potentially causes the case where we return a null reference.