Adding constructor detection for speculative output types
If an `AutoFilter` signature contains a by-reference output type of the form `T&`, and `T(AutoPacket&)` can be constructed, then this construction signature will be invoked instead of the default constructor, even when no default constructor exists.  This provides output decorations with a way of obtaining the current packet during construction in order to provide a superior strategy for implementing new AutoFilter argument behaviors.