Altitude is now feature complete
This adds the meta-functions necessary to extract an altitude from a class, the modifications of PulseSatisfaction to make DecorateImmediate work, an overload of AutoPacketFactory's += to allow a altitude specified filter to be added ad-hoc more easily, and several more unit tests for altitude.
At this point the unit test coverage of the feature should be reasonable, but it needs to be brought up even higher in order to be sure that the code all works as expected.