Deprecate ContextCreator
This type confusingly conflates the idea of context creation with context mapping.  A separate type, ContextMap, already exists to function as a map between a key and a context, and should be preferred.  Users should make use of `CoreContext::Create` or `AutoCreateContext` if they intend to create and manage a keyed set of contexts.