Reintroducing CoreContext::Add
This method allows users to explicitly register a type in a context for discovery without making that type fully present in the context as a concrete type.  The specified interface will not be the subject of event receipt nor will it participate in `AutoFilter` networks.  `CoreContext::Add` is the logical counterpart to `CoreContext::Snoop`.