Completion status moved into its own status block
In the single context owner case, when the sole external holder of a `CoreContext` invokes `CoreContext::Wait`, that owner must hold the sole exclusive shared pointer to the context on return.  This means that threads running in that context must not transition themselves to the "completed" state until after releasing all shared pointers to their enclosing context.

In order to do this, the "completed" status flag has been moved to the context block, in order to allow us to update it even in the "lights out" countercase where the last holder of the context is a runnable in that same context.