Eliminate libstdc++ support from CMakeLists
The time has finally come to eliminate support for libstdc++ once and for all.  We will start by short-circuiting all preprocessor blocks and macros that make references to libcxx.  Then comes the slow, arduous process of fixing all of the stupid backflips that we had to make to classes such as `ObjectPool` to make them compatible with libstdc++.

libstdc++ is dead.  Long live libc++.