Recursively call `CoreContext::Stop` before waiting on anything
This change ensures that all members of the current context receive an `OnStop` message before a wait is triggered on child contexts.

It's also gradually becoming clear that it may be necessary to deprecate `SignalShutdown(true)` as a typical calling convention.  We should probably require users to manually call `Wait` if that's what they want to do.