AutoNetServerImpl must stop itself when an exception is thrown by CoreThread::Run
We have an `std::future` in scope at this point.  This future will block and wait for the transport to terminate, which won't happen until the context is stopped.  This will prevent the `CoreThread::Run` unhandled exception handler from ever receiving control.  To prevent deadlock, this condition must be specifically addressed and prevented.