Enhance documentation and fix teardown listener behavior for BasicThread
Add documentation to describe exactly when teardown listeners are raised for `BasicThread`.  Presently, the documentation for this method leaves things rather open-ended.  Also change NotifyTeardownListeners so it is invoked when the owning type is still a `BasicThread`, and only invoke `NotifyTeardownListeners` after the `BasicThread` has transitioned to the stopped state.