Add an altitude concept to describe filter priorities
Altitude is now declared on an AutoFilter as a static member of that class, as so:

```C++
class MyFilter {
    static const autowiring::altitude altitude = autowiring::altitude::Instrumentation;
};
```

The remainder of the AutoFilter declaration is unchanged.  By default, AutoFilters marked `Deferred` will run at a higher altitude than other filters.

Currently, altitudes can only be easily declared on class-level AutoFilters.