Refactor DispatchQueue to use an intrusive list instead of an std::list
An intrusive linked list, in this case, is a much better choice.  Heap allocations for each entry in the collection are unavoidably necessary because entries have a variable size, which means a linked list is an improvement over `std::list`.

If necessary, a next step could be to use a journal-based arena allocator in order to keep links in the list sequential with respect to each other.