Fix bug in AnySharedPointer::operator<
This bug is being caused by an implicit cast from `SharedPointerSlot` to `bool`.  The resulting boolean is being used in the various comparison operations silently.  Mark `SharedPointerSlot::operator bool` as `explicit` and then correct the compiler errors that result.

This bug is causing AutoFilters to sometimes be considered equal even when they are not.  It manifests most obviously when multiple AutoFilter lambdas are being attached to the same packet, but have different pointer values.  Currently, the most common use case of `AutoPacketFactory::operator+=` only ever attaches one instance of an AutoFilter lambda, but if multiple similar instances are attached, or if the linker performs COMDAT reduction and causes two distinct lambdas to share the same code, an error will result.

Also add a test to guard against a regression here, both in the comparators and in the above folding case.