Correct race condition in auto_prev
Rename `AddSatCounter` to `AddSatCounterUnsafe`, because this method is unsynchronized.  This oversight created a race condition in `AutoPacketInternal::Initialize`; formerly, this method was called before the packet was ever exposed to the caller.  Now, the method may potentially be invoked while a filter is already running.

Add a pathological test to verify this behavior by making use of the same `AutoPacket` in two thread contexts.