Add a ThreadPool concept for CoreContext
The thread pool concept allows users to attach a lambda to the context to be executed by that context's thread pool.  By default, contexts receive the thread pool of their parent, and the global context gets the default thread pool for the system.  Contexts assign their thread pool to the thread pool of their parent upon initiation (not construction).  Prior to initiation, contexts hold the null thread pool, which is an instance of NullPool that simply accumulates dispatchers until the parent context is initiated, shut down, or destroyed.

Setting a context's thread pool will not cause the context to release its reference to the NullPool.  Instead, the provided thread pool will be held until the context is started, and when this happens, all dispatchers will be transferred to the requested pool and then the requested pool will be initiated.

Jobs added to a context's thread pool will not be run until the context is started.