Create DispatchThunk outside of lock
This creation step could potentially cause reentrancy, even though it's supposed to be a relatively safe `std::move` of what is presumably a lambda function.  Move the create call outside of the lock in order to eliminate this as a possibility and also slightly reduce the amount of time that we spend holding a lock.