Fix an exception thrown when DecorateImmediate is used more than once
The exception is thrown when the second call to `AutoPacket::DecorateImmediate` would have satisfied the AutoFilter if the first call were made with `AutoPacket::Decorate` instead.  `AutoPacket` is not correctly detecting that the first call was transient, and is incorrectly attempting to make the AutoFilter call at the point of the second `Decorate` call.