Implement BasicThread::IsMainThread
Add support to discriminate between the main thread and other threads on the system.  Some systems have special handling for the main thread--IE, certain APIs can only be called safely from it--thus it's important that we have a way to inform such systems as to whether a particular thread is the main thread.