Do not use isspace on non-ASCII strings
Characters like ├ are not actually ASCII characters, and it is an error to pass them to routines like `isspace`.  On Windows, this actually causes an assertion failure.  Furthermore, this test should probably care about the specific placement of newlines.

Use a handrolled lambda to eliminate spaces in particular while preserving newlines to get around this issue.