Implement auto_out
`auto_out` is a feature that allows filters to defer the satisfaction of their output arguments until after they return.  `auto_out` implements `shared_ptr` semantics in that it is movable and copyable, and therefore allows satisfaction responsibility to be transferred or shared.

This feature will eventually be used to support a simplified fan-out concept.