Add representational equivalence check
A planned optimization involves minimizing the use of `AnySharedPointer` in favor of `std::shared_ptr<void>`.  To make this change, it will be necessary to ensure that this type can be reinterpret casted to `std::shared_ptr<T>` for all `T`.

Add a unit test to ensure that this cast is possible and well-behaved on all target platforms.