Fix issue where unit test is allowing state leakage
This unit test uses the global context, and so is allowing leakage of state in cases where contexts do not cleanly tear down in other tests.  Fix it so it references only those elements created locally in order to prevent spurious errors from being reported here.  Also add an override to `PrintContextTree` that accepts a specific `CoreContext` whose tree is to be printed.