Minor optimization for zero-duration delayed dispatchers
If a delayed lambda is pended to a `DispatchQueue` with a delay duration of zero, then just pend this to the ready queue directly, completely bypassing the waiting queue.

Also, change the time offset calculation for relatively delayed lambdas to be performed after the lambda is constructed and about to be pended to the queue.