Implement a heterogenous block concept
Heterogenous blocks are intended to be used by AutoPacket.  They contain a multitude of unrelated types that are packed together in the same arena based allocator.  The intended use case is for situations where a particular decoration is known to never be needed as a `shared_ptr`.  Such cases free us from having to prospectively allocate a shared pointer.

The heterogenous block will also need to be designed to support post-hoc generation of a shared pointer.  This is necessary because there will be a minority of cases where a shared pointer is unexpectedly needed.