CoreContext must not be abstract
There are cases where `CoreContext` methods are invoked on instances of this type while `CoreContext` is in teardown.  Generally this safe from a state standpoint, but we must be sure that none of those methods are pure virtual or invoke any pure virtual methods.

Make `CoreContext` a concrete type, and add a static assertion in the `CoreContext` unit test to ensure that the type does not become abstract again in the future.