Rework internal synchronization mechanism in autowiring::signal
Fix a race condition where a crash can occur if signal registration were to happen asynchronously with respect to signal assertion.  To fix this, a stronger guarantee has to be made that the only threads which will invoke listeners will be those threads which assert the signal.  Signal registration and unregistration must not invoke listeners.

This does mean, however, that up to one deferred call entry may remain stuck to the signal in a zombie state until the time of the next call, and if the next call never happens, this could be indefinite.  Care must be taken to ensure that the underlying listener itself is properly destroyed.  Add a unit test to assure this.