Use case improvements for autowiring::parallel
This type is a little hard to use because the underlying thread pool is unpredictable.  Typical use cases do not mind the construction expense of this type, and instead simply prefer that the behavior is predictable.  Therefore, optimize on these use cases, and manually maintain a thread pool for system use.