Deprecate `CoreContext` thread pools
This concept is hard to manage and has very unexpected behavior.  Contexts should be containers used to manage a fixed set of well-defined classes; the thread pool should be used as a utility type for handling pooling operations, and the two types should not really meet.

At a later time it might make sense to have `ThreadPool` inherit `CoreRunnable`, but for now it's fine to keep the interfaces unrelated.