Add a configuration concept to Autowiring
To start, this concept is self-contained and shows a way to set a string value by a string name, and how to define a descriptor on a type.  Eventually we will make contexts aware of this concept, and build out a way to set all configuration values found in a context to their respective values.

This is the initial work to implement this feature.  Metadata and scanner concepts still needs to be introduced.