Eliminate AutoFired
This concept has finally been removed from all downstream users, and should be eliminated everywhere.  `AutoFired` is a legacy concept and the many-to-many relationship that it enables between components tends to create more maintenance difficulties than it resolves.

Eliminate `AutoFired` and all of its attendant functionality.  Encourage downstream consumers to switch to signals.