Retroactive bolting for existing contexts
Historically, bolts are only asserted for _new_ contexts which match the bolt pattern.  This should be changed so that a bolt will be asserted for any context which already exists at the time the bolt is introduced.

When a bolt is added to a context, it should be asserted for all parents and children that are already created which satisfy the bolt requirement.  This allows bolts to be added potentially at some point after a context is created and running, and allows a new order-independent pattern to be considered the standard behavior in broader parts of the system.

Also change the bolt system so that it internally uses `auto_id` instead, this type is faster and more flexible than the STL `type_info`, and fix the hashing routine so that `auto_id{}` may be hashed.