Add tid to BasicThread, make debug lists more readable.
Another slight adjustment to hung runnable outputs.  Contexts should print all of their runnable members, whether they are running or not, and use a status flag to indicate what is still running and under what context.  Also print the thread ID of runnables that are threads, wherever possible.

Furthermore, the context map was not actually readable when printed in windows due to Unicode translation problems on the command line.  Replace the unicode box drawing code with simple ASCII in order to ensure compatibility for now, and update unit tests as needed.

Also add a unit test for all new behaviors.