Fix leak in thread_specific_ptr
This leak is a slow leak on Windows caused by the way that [`TlsSetValue`](https://msdn.microsoft.com/en-us/library/windows/desktop/ms686818(v=vs.85).aspx) works.  If the current context is assigned for a thread and never unassigned, then the shared pointer allocated for that thread could potentially not be freed, resulting in the leak.

Fix this issue by using [`FlsSetValue`](https://msdn.microsoft.com/en-us/library/windows/desktop/ms683146(v=vs.85).aspx) instead, which allows a teardown routine to be specified.  On platforms that do not support this function, fall back to the old behavior.