Invert AutoFilter altitudes
The original concept of an AutoFilter altitude causes the filter with the highest altitude value to be invoked first.  Unfortunately, this means that an enumeration defining the altitudes for use in a filter network must be declared in the reverse order of invocation, which is confusing.

Flip the altitude test around, and make lower filters run before higher filters.